Use separate packages for examples
