--- examples/lambda-user/src/main/java/info/javaspec/client/Player.java
@@ -1,3 +1,0 @@
-package info.javaspec.client;
-
-class Player {}
