+++ examples/lambda-user/src/main/java/info/javaspec/example/greeter/Greeter.java
@@ -1,4 +1,4 @@
-package info.javaspec.client;
+package info.javaspec.example.greeter;
 
 public class Greeter {
 	public String makeGreeting() {
