+++ examples/lambda-user/src/main/java/info/javaspec/minimax/Minimax.java
@@ -1,4 +1,4 @@
-package info.javaspec.client;
+package info.javaspec.example.minimax;
 
 class Minimax {
 	private final Player maximizer;
