+++ examples/lambda-user/src/main/java/info/javaspec/minimax/Player.java
@@ -1,0 +1,4 @@
+package info.javaspec.example.minimax;
+
+class Player {
+}
