+++ examples/lambda-user/src/test/java/info/javaspec/example/greeter/GreeterSpecs.java
@@ -1,4 +1,4 @@
-package info.javaspec.client;
+package info.javaspec.example.greeter;
 
 import static org.junit.jupiter.api.Assertions.*;
