+++ examples/lambda-user/src/test/java/info/javaspec/example/minimax/MinimaxSpecs.java
@@ -1,4 +1,4 @@
-package info.javaspec.client;
+package info.javaspec.example.minimax;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
